The cell responsible for displaying each element in Settings List.
